Compilation errors fixed for environments, where qreal is not a double
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk@886 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ qwt/src/qwt_interval_symbol.cpp
@@ -181,7 +181,7 @@ const QPen& QwtIntervalSymbol::pen() const
 void QwtIntervalSymbol::draw( QPainter *painter, Qt::Orientation orientation,
     const QPointF &from, const QPointF &to ) const
 {
-    const double pw = qMax( painter->pen().widthF(), 1.0 );
+    const qreal pw = qMax( painter->pen().widthF(), qreal( 1.0 ) );
 
     QPointF p1 = from;
     QPointF p2 = to;
